fix(management): Better handling of settings
Closes gravitee-io/issues#4787
