+++ gravitee-rest-api-management/gravitee-rest-api-management-rest/src/main/java/io/gravitee/rest/api/management/rest/resource/EnvironmentResource.java
@@ -265,4 +265,9 @@ public class EnvironmentResource extends AbstractResource {
     public NotifiersResource getNotifiersResource() {
         return resourceContext.getResource(NotifiersResource.class);
     }
+
+    @Path("settings")
+    public SettingsResource getSettingsResource() {
+        return resourceContext.getResource(SettingsResource.class);
+    }
 }
