+++ gravitee-rest-api-model/src/main/java/io/gravitee/rest/api/model/config/Analytics.java
@@ -13,18 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service;
+package io.gravitee.rest.api.model.config;
 
-import io.gravitee.rest.api.model.config.ConsoleConfigEntity;
-import io.gravitee.rest.api.model.config.PortalConfigEntity;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import io.gravitee.rest.api.model.annotations.ParameterKey;
+import io.gravitee.rest.api.model.parameters.Key;
 
-/**
- * @author Nicolas GERAUD (nicolas.geraud at graviteesource.com) 
- * @author GraviteeSource Team
- */
-public interface ConfigService {
-    ConsoleConfigEntity getConsoleConfig();
-    PortalConfigEntity getPortalConfig();
-    boolean portalLoginForced();
-    void save(ConsoleConfigEntity consoleConfigEntity);
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Analytics {
+    @ParameterKey(Key.ANALYTICS_CLIENT_TIMEOUT)
+    private Long clientTimeout;
+
+    public Long getClientTimeout() {
+        return clientTimeout;
+    }
+
+    public void setClientTimeout(Long clientTimeout) {
+        this.clientTimeout = clientTimeout;
+    }
 }
