+++ gravitee-rest-api-model/src/main/java/io/gravitee/rest/api/model/config/ReCaptcha.java
@@ -13,21 +13,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.model.annotations;
+package io.gravitee.rest.api.model.config;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import io.gravitee.rest.api.model.annotations.ParameterKey;
 import io.gravitee.rest.api.model.parameters.Key;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ReCaptcha {
+    @ParameterKey(Key.RECAPTCHA_ENABLED)
+    private Boolean enabled;
+    @ParameterKey(Key.RECAPTCHA_SITE_KEY)
+    private String siteKey;
 
-/**
- * @author Nicolas GERAUD (nicolas.geraud at graviteesource.com)
- * @author GraviteeSource Team
- */
-@Target({ElementType.FIELD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ParameterKey {
-   Key value();
+    public Boolean getEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(Boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getSiteKey() {
+        return siteKey;
+    }
+
+    public void setSiteKey(String siteKey) {
+        this.siteKey = siteKey;
+    }
 }
