+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/configuration/application/ApplicationTypeServiceImpl.java
@@ -17,18 +17,14 @@ package io.gravitee.rest.api.service.impl.configuration.application;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.gravitee.repository.exceptions.TechnicalException;
-import io.gravitee.rest.api.model.PortalConfigEntity;
+import io.gravitee.rest.api.model.config.Application;
 import io.gravitee.rest.api.model.configuration.application.ApplicationTypeEntity;
 import io.gravitee.rest.api.model.configuration.application.ApplicationTypesEntity;
 import io.gravitee.rest.api.service.ConfigService;
 import io.gravitee.rest.api.service.configuration.application.ApplicationTypeService;
 import io.gravitee.rest.api.service.exceptions.ApplicationTypeNotFoundException;
 import io.gravitee.rest.api.service.exceptions.TechnicalManagementException;
-import io.gravitee.rest.api.service.impl.ApplicationServiceImpl;
 import org.apache.commons.io.IOUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -93,8 +89,8 @@ public class ApplicationTypeServiceImpl implements ApplicationTypeService {
     }
 
     public JsonNode getApplicationTypesConfiguration() {
-        PortalConfigEntity.Application applicationConfig = configService.getPortalConfig().getApplication();
-        PortalConfigEntity.Application.ApplicationTypes types = applicationConfig.getTypes();
+        Application applicationConfig = configService.getConsoleConfig().getApplication();
+        Application.ApplicationTypes types = applicationConfig.getTypes();
         if (!applicationConfig.getRegistration().getEnabled()) {
             types.getBrowserType().setEnabled(false);
             types.getBackendToBackendType().setEnabled(false);
